Convert this ghay code to maven
Bruh maven works wonders
