+++ .gitignore
@@ -1,0 +1,5 @@
+/target/
+.classpath
+/.idea/
+.project
+/.settings/
