--- .idea/libraries/nukkit_1_0_SNAPSHOT.xml
@@ -1,9 +1,0 @@
-<component name="libraryTable">
-  <library name="nukkit-1.0-SNAPSHOT">
-    <CLASSES>
-      <root url="jar://$USER_HOME$/Downloads/Real Nukkit/nukkit-1.0-SNAPSHOT.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
